Commit 1
Adopted the AutoUpdater script for the Chrono. Began reading through the
code to learn how the program worked. After learning how it works,
noticed that the live version is having errors with Google Sheets after
a while of running the script.

It appears that red messages are not showing up where expected in the
code. When they do happen, they throw off the rest of the script's
proceedure because it's checking for colors in the exact spot where they
will not show up. Will need to adjust where it is looking, save the
script, rerun and monitor for errors.