Commit 10
Making maximize window happen only once.

The script would maximize window elements as part of it's code, but it would only need to do this once, not every time the script loops. When it happens, it causes warning popups to be large and harder to deal with when the script stops working.