Refinining Updater for Backlog updating.
While testing the operation, I found that it wouldn't be useful to be getting the array's in the updater and comparing because they'll both be very different.

Instead, I'm going to put the backlog into a staging area and work at formatting it to match the target and then comparing for update.