Fixed Ally to do assignments in order, updated eslint
Ally was not assigning accounts for Outsource designers in the order that was presented in the backlog.

Removing sorting from Ally fixes this issue, but this also affected how designers were able to update an account's status.

Removing backlog sorting from all places where it was referenced throughout Ally fixes the issue and now Ally is working fine.